JWT Authentication

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler')
 
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 
 //@desc     Get all goals
 //@route    GET /api/goals
 //@access   Private
 const getGoals = asyncHandler(async(req, res) => {
-    const goals = await Goal.find({})
+    const goals = await Goal.find({user: req.user._id})
     res.status(200).json(goals)
 })
 
@@ -20,7 +21,8 @@ const setGoal = asyncHandler(async(req, res) => {
     }
 
     const goal = await Goal.create({
-        text: req.body.text
+        text: req.body.text,
+        user: req.user._id
     })
 
     res.status(200).json(goal)
@@ -35,6 +37,18 @@ const updateGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(404)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user._id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== req.user._id.toString()){
+        res.status(401)
+        throw new Error('You are not authorized to update this goal')
     }
 
     const updateGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -53,6 +67,18 @@ const deleteGoal = asyncHandler(async(req, res) => {
     if(!goal){
         res.status(404)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user._id)
+
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    if(goal.user.toString() !== req.user._id.toString()){
+        res.status(401)
+        throw new Error('You are not authorized to delete this goal')
     }
 
     await Goal.findByIdAndDelete(req.params.id)
